+++ praetorian_ssh_proxy/hanlers/menu_handler.py
@@ -2,25 +2,26 @@ import sys
 
 
 class MenuHandler(object):
-    def __init__(self, client, client_channel):
+    def __init__(self, client, client_channel, remote_checker):
         self._client = client
         self._client_channel = client_channel
         self._buffer = ''
+        self._remote_checker = remote_checker
 
     @staticmethod
-    def create_from_channel(client, client_channel) -> 'MenuHandler':
-        return MenuHandler(client, client_channel)
+    def create_from_channel(client, client_channel, remote_checker) -> 'MenuHandler':
+        return MenuHandler(client, client_channel, remote_checker)
 
-    def serve_remote_menu(self, remotes):
-        if isinstance(remotes, list):
+    def serve_remote_menu(self):
+        if not self._remote_checker.is_remote_set:
             self._client_channel.send('------------------------------------\r\n')
             self._client_channel.send('| Welcome to Praetorian SSH proxy  |\r\n')
             self._client_channel.send('------------------------------------\r\n')
-            for counter, remote in enumerate(remotes, 1):
+            for counter, remote in enumerate(self._remote_checker.remote, 1):
                 self._client_channel.send('| {:30} {} |\r\n'.format(remote.name, counter))
 
             self._client_channel.send('------------------------------------\r\n')
-            self._client_channel.send('| {:30} {} |\r\n'.format('exit', len(remotes) + 1))
+            self._client_channel.send('| {:30} {} |\r\n'.format('exit', len(self._remote_checker.remote) + 1))
             self._client_channel.send('------------------------------------\r\n')
             self._client_channel.send('Choose your remote: ')
 
@@ -43,12 +44,12 @@ class MenuHandler(object):
 
                 # ENTER
                 elif data == b'\r':
-                    if self._buffer in map(str, range(1, len(remotes) + 1)):
-                        remote = remotes[int(self._buffer) - 1]
-                        self._client_channel.send(f'\n\rChosen remote {remote.name}.\r\n')
-                        return remote
+                    if self._buffer in map(str, range(1, len(self._remote_checker.remote) + 1)):
+                        self._remote_checker.set_remote(self._remote_checker.remote[int(self._buffer) - 1])
+                        self._client_channel.send(f'\n\rChosen remote {self._remote_checker.remote.name}.\r\n')
+                        return
 
-                    elif self._buffer == str(len(remotes) + 1):
+                    elif self._buffer == str(len(self._remote_checker.remote) + 1):
                         self._client_channel.send(f'\n\rExiting ...\r\n')
                         self._client.close()
                         sys.exit(0)