+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 - 🥝 Implemented interactive shell session.
 - 🥝 Implemented support for fabric script execution.
 - 🥝 Implemented connection and integration with praetorian-api.
+
+## 0.3.0 : 2020-12-12
+
+- 🥐 Implemented Remote object checker
+- 🥐 Refactored connection and authentication workflow.